requirements.txt
astroid==2.2.5
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pkg-resources==0.0.0
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
